files + math stuff working
