+++ package.json
@@ -46,8 +46,7 @@
     "@electron-forge/maker-zip": "^6.0.0-beta.30",
     "assert": "^1.4.1",
     "electron": "3.0.10",
-    "electron-rebuild": "^1.8.2",
-    "mocha": "^5.2.0"
+    "electron-rebuild": "^1.8.2"
   },
   "eslintConfig": {
     "env": {
